+++ IBActionSheetSample/IBActionSheetSample/ViewController.m
@@ -128,7 +128,7 @@
 
 // the delegate method to receive notifications is exactly the same as the one for UIActionSheet
 - (void)actionSheet:(IBActionSheet *)actionSheet clickedButtonAtIndex:(NSInteger)buttonIndex {
-    NSLog(@"Button at index: %d clicked\nIt's title is '%@'", buttonIndex, [actionSheet buttonTitleAtIndex:buttonIndex]);
+    NSLog(@"Button at index: %ld clicked\nIt's title is '%@'", (long)buttonIndex, [actionSheet buttonTitleAtIndex:buttonIndex]);
 }
